feat: Added in checking/offering of a player redirection
If the player in use is mediastore, there is a chance that
the media can be played back on mediaplayer. This
code checks to see whether that is a potential, and if
so it will offer the user the chance to try out the
redirection to mediaplayer. The HTML/CSS of the exta
additions to the mediastore page beneath the video
that offer the redirection really do need a great deal
of tidying up, but most of the desired functionality
is there.
Next steps are to add a plugin-wide default of
whether to offer the chance to redirect, or to just
try it automatically if it appears that it may be
valid to do.