refactor: Added in check for Mediaplayer
Now that the plugin is checking for the
Mediastore and Mediaplayer pages, we can
no longer blindly apply the changes to the
page, assuming that it is a Mediaplayer one.
This just adds in a couple of checks to make
sure the player has been identified as the
Mediaplayer one, before trying to do
Mediaplayer stuff on it.