feat: Working around Mediaplayers autoplay
As the watch party will require the users to enable autoplay, that does
not mean we want Mediaplayer to be the one doing the autoplaying.
So this just adds a check to make sure the initial play command to the
video is either from a user's interactions with the player, or from the
watch party code.