feat: Added Default Resolution Radios to the HTML
Instead of trying to load this dynamically at runtime, it is much easier
to just code these into the HTML for the settings menu instead.
The ID of the setting needs to match the setting name in the JS, as well
as the name of the radio buttons in the HTML, for the javascript to
work correctly. The radio buttons need to be held in something with the
radioHolder type, and they need to have the class radioHolderEntry.